move instrument specific things to separate module
